code/requirements.txt
fastapi[all]==0.128.0
uvicorn==0.40.0
SQLAlchemy[asyncio]==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.11
python-multipart==0.0.22
pydantic==2.10.6
python-dotenv==1.0.1
python-magic==0.4.27
aiofiles==25.1.0
pillow==12.1.0
aiohttp==3.13.3
requests==2.32.5